pyproject.toml
[project]
name = "fastapi-deprecation"
dynamic = ["version"]
description = "Endpoint deprecation management for FastAPI made easy"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "fastapi>=0.129.0",
    "pydantic>=2.12.5",
    "python-dateutil>=2.9.0.post0",
    "typing-extensions>=4.15.0",
]

[build-system]
requires = ["hatchling", "hatch-vcs"]
build-backend = "hatchling.build"

[tool.hatch.version]
source = "vcs"

[tool.hatch.build.hooks.vcs]
version-file = "src/fastapi_deprecation/_version.py"

[dependency-groups]
dev = [
    "coverage>=7.13.4",
    "httpx>=0.28.1",
    "mkdocs>=1.6.1",
    "mkdocs-material>=9.7.2",
    "mkdocstrings[python]>=1.0.3",
    "pre-commit>=3.5.0",
    "pytest>=9.0.2",
    "pytest-cov>=7.0.0",
    "ruff>=0.15.1",
]
